feat: feat/add-docker-reg-secret
